+++ packages/highlight/src/themes/custom/one/light.ts
@@ -279,6 +279,20 @@ const oneLight = new HighlightCustomTheme(
         fontStyle: "italic",
       },
     },
+    {
+      types: ["banners"],
+      languages: ["myBanner"], 
+      style: {
+        color: "hsl(286, 60%, 67%)"
+      }
+    },
+    {
+      types: ["banner-end"],
+      languages: ["myBanner"], 
+      style: {
+        color: "hsl(133, 45%, 53%)"
+      }
+    }
   ]
 );